refactor: TodayContent 공통 스타일 파일로 분리
