+++ client/src/features/TodayMoment/ui/TodayContent.styles.ts
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 
-export const TodayMomentContentWrapper = styled.div`
+export const TodayContentWrapper = styled.div`
   display: flex;
   flex-direction: column;
   gap: 20px;
