+++ client/src/features/TodayMoment/ui/TodayMomentWriteContent.tsx
@@ -2,7 +2,7 @@ import { YellowSquareButton } from '@/shared/ui/button/YellowSquareButton';
 import { Card, TextArea } from '@/shared/ui';
 import { Send, Star } from 'lucide-react';
 import { useState } from 'react';
-import * as S from './TodayMomentContent.styles';
+import * as S from './TodayContent.styles';
 
 interface TodayMomentWriteContentProps {
   onSubmit: () => void;
@@ -23,7 +23,7 @@ export const TodayMomentWriteContent = ({ onSubmit }: TodayMomentWriteContentPro
     onSubmit();
   };
   return (
-    <S.TodayMomentContentWrapper>
+    <S.TodayContentWrapper>
       <Card.TitleContainer
         Icon={Star}
         title="모멘트 공유하기"
@@ -48,6 +48,6 @@ export const TodayMomentWriteContent = ({ onSubmit }: TodayMomentWriteContentPro
           disabled={text.trim().length === 0}
         />
       </Card.Action>
-    </S.TodayMomentContentWrapper>
+    </S.TodayContentWrapper>
   );
 };
